Refactor and setup.py
