+++ pixelart_tui/command_line.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+from pixelart_tui.main import SimpleApp
+
+def run():
+    SimpleApp.run()
+
