+++ pixelart_tui/my_widgets.py
@@ -23,10 +23,10 @@ from rich.console import (
     RenderResult,
     RenderableType,
 )
-from my_button import MyButton
+from pixelart_tui.my_button import MyButton
 from typing import List
 
-from my_messages import DebugStatus, ColorStatus, LoadSaveStatus
+from pixelart_tui.my_messages import DebugStatus, ColorStatus, LoadSaveStatus
 
 def minmax(a, mn, mx):
     return max(min(mx, a), mn)
