+++ run.bat
@@ -1,1 +1,1 @@
-python3 main.py
+python3 .\pixelart_tui\main.py
