+++ run.sh
@@ -1,1 +1,1 @@
-python3 text_game.py
+python3 ./pixelart_tui/text_game.py
